+++ packages/kit/src/views/Market/MarketDetailV2/components/InformationTabs/layout/DesktopInformationTabs.tsx
@@ -1,6 +1,8 @@
+import { useCallback, useMemo } from 'react';
+
 import { useIntl } from 'react-intl';
 
-import { Tabs } from '@onekeyhq/components';
+import { Tabs, YStack } from '@onekeyhq/components';
 import { getNetworkIdsMap } from '@onekeyhq/shared/src/config/networkIds';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
 
@@ -8,17 +10,37 @@ import { useTokenDetail } from '../../../hooks/useTokenDetail';
 import { Holders } from '../components/Holders';
 import { TransactionsHistory } from '../components/TransactionsHistory';
 
+import type { TabBarProps } from 'react-native-collapsible-tab-view';
+import { StickyHeader } from './StickyHeader';
+
+function DesktopInformationTabsHeader(props: TabBarProps<string>) {
+  const firstTabName = useMemo(() => {
+    const { tabNames } = props;
+    return tabNames[0];
+  }, [props]);
+  return (
+    <YStack bg="$bgApp" pointerEvents="box-none">
+      <Tabs.TabBar {...props} />
+      <StickyHeader firstTabName={firstTabName} />
+    </YStack>
+  );
+}
+
 export function DesktopInformationTabs() {
   const intl = useIntl();
   const { tokenAddress, networkId } = useTokenDetail();
   const networkIdsMap = getNetworkIdsMap();
+
+  const renderTabBar = useCallback(({ ...props }: any) => {
+    return <DesktopInformationTabsHeader {...props} />;
+  }, []);
 
   if (!tokenAddress || !networkId) {
     return null;
   }
 
   return (
-    <Tabs.Container renderTabBar={(props) => <Tabs.TabBar {...props} />}>
+    <Tabs.Container renderTabBar={renderTabBar}>
       <Tabs.Tab
         name={intl.formatMessage({
           id: ETranslations.dexmarket_details_transactions,
